requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.87.0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-decouple==3.6
sniffio==1.3.0
starlette==0.21.0
typing-extensions==4.4.0
uvicorn==0.19.0
zipp==3.10.0
